Move common API code to common.
